Add files via upload
A program that accepts the lengths of three sides of a triangle as inputs. The program should output whether or not the triangle is a right triangle (Recall from the Pythagorean Theorem that in a right triangle, the square of one side equals the sum of the squares of the other two sides). Implement using functions.

Example Runs:

Example 1

    Enter the length of the first side: 3
    Enter the length of the second side: 4
    Enter the length of the third side: 5
    The given triangle is a right triangle.

Example 2:

    Enter the length of the first side: 2
    Enter the length of the second side: 3
    Enter the length of the third side: 4
    The given triangle is not a right triangle.